Bugfix: Sometimes CalcBucketPos has three bucket layers (#48)
# Description
- Fixed issue that vertical height start below the grid => lower all z
index by one
- Fixed issue when three bucket layers are detected

# Note
The presence of three bucket layers seems to not be a bug but expected
considering the current implementation.
The issue is that a cell is considered as a bucket cell even if only a
tiny part of the bucket is included. So for some specific inclination of
the bucket, it may cause issue close to a vertex, in particular close to
the teeth when the bucket is almost vertical.
The new implementation has been tested with MuJoCo, and the results show
that every time there are three bucket layers, there is a maximum gap
between two bucket layer of 1.5 the maximum cell size, which is fairly
low.